interim update, adding ping status page, DAQ-less water only, fix for runtime calculations when there are no existing mappings

+++ xeros-ldcs/src/main/java/com/elyxor/xeros/ldcs/dai/WaterMeterPortInterface.java
@@ -7,7 +7,7 @@ public interface WaterMeterPortInterface {
 	public long[] getPrevMeters();
 
 	public long getWaterMeterId();
-	public void setWaterMeterId(int id);
+	public void setWaterMeterId(long id);
 
 	//commands
 	public String initRequest();
